2024Jul04-1
HQP: Added new rules to remove blank square when videos are paused on mobile, added rule to remove black friday banner (when applicable), removes a banner.js from the players website, removes an ad container and lastly blocks the hf page that shows when a video is paused. Also changed a rule to be a wildcard rule for smartpop instead of focusing on an old specific ad domain.