2024Nov24-1
FikFap: Changes to now show that a live broadcast /would/ have been shown by not removing the LIVE at the bottom and allowing the API to reply with a valid HLS url, which is blocked, resulting in a loading circle. It's a better solution than before.
Xfree: Updated rules to remove new annoyances and tracking.