2025Sep07-2
Bunkr: Fixed an issue where advanced album view didn't load properly due to an old rule.
Watchdirty: A couple of small fixes of previous rules that weren't set to use any TLD, and added rules to hopefully block navbar ads and friends grid in the future. Also added a rule to block to ahrefs for privacy.